Show cache info (#54)
* Fix cached node colour

* Add function to get run ID for execution

* Store cache configs in separate table

* Use model when loading steps for run

* Indicate when written to cache

* Fix converting cache expiration to milliseconds

* Show information about step's cache configuration

* Tidy/fix client-side types

* Use blob type for cache/etc keys

+++ clients/python/coflux/models.py
@@ -15,7 +15,7 @@ class Parameter(t.NamedTuple):
 
 class Cache(t.NamedTuple):
     params: list[int] | t.Literal[True]
-    max_age: float | None
+    max_age: int | None
     namespace: str | None
     version: str | None
